优化字段拓展
