+++ hsweb-starter/hsweb-spring-boot-starter/src/main/java/org/hswebframework/web/starter/entity/EntityMappingCustomer.java
@@ -1,0 +1,12 @@
+package org.hswebframework.web.starter.entity;
+
+import org.hswebframework.web.commons.entity.factory.MapperEntityFactory;
+
+/**
+ * TODO 完成注释
+ *
+ * @author zhouhao
+ */
+public interface EntityMappingCustomer {
+    void customize(MapperEntityFactory entityFactory);
+}
